cleaned up code: some comiler-warnings fixed, obsolete code removed (oops..and some whitespace stuff) (by CleaningWoman)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6102 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Handle UI.cpp
@@ -1284,7 +1284,7 @@ UINT32 UIHandleEndTurn( UI_EVENT *pUIEvent )
 		 // Just try/catch (ugly, but works).	 
 		 __try
 		 {		 
-			 for(UINT32 i=0; i<WORLD_MAX ;i+=4) 
+			 for(UINT32 i=0; i<(UINT32)WORLD_MAX ;i+=4) 
 			 {			
 				gubWorldTileInLight[i] = InLightAtNight(i, gpWorldLevelData[ i ].sHeight);
 				gubIsCorpseThere[i] = IsCorpseAtGridNo( i, gpWorldLevelData[ i ].sHeight );
